Finished main part of project
It now functions as a perceptron when configured to find the average of 2 input.

It took me a while to realize I was using the wrong step size, so I fixed that and want to push my code before I find out if the additional functionality beyond a perceptron works/fix it if it doesn't.