Got it to run on my laptop
Removed and replaced all files after creating a new project by the same name in Pycharm.
As far as I could tell, this was the only way to create a new venv instance and interpreter for the project.
All my other interpreters seem to be invalidated or corrupted for some reason.